requirements.txt
click
flask
flask-shell-ipython
mock
pytest
pyzmq
tox
